+++ test/z_close.js
@@ -1,0 +1,6 @@
+var server = require('./server');
+
+module.exports.server = function(assert) {
+    server.close();
+    assert.done();
+}